Validations added on menu-item --NOT WORKING ON PARTIAL OBJECTS

+++ src/menu-item/menu-item.controller.ts
@@ -7,10 +7,13 @@ import {
   Param,
   Body,
   UsePipes,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { MenuItemDTO } from './menu-item.dto';
 import { MenuItemService } from './menu-item.service';
 import { ValidationPipe } from '../shared/validation.pipe';
+var validate = require('uuid-validate');
 
 @Controller('menu-item')
 export class MenuItemController {
@@ -23,6 +26,10 @@ export class MenuItemController {
 
   @Get(':id')
   getMenuItemById(@Param('id') id: string) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemService.read(id);
   }
 
@@ -35,11 +42,19 @@ export class MenuItemController {
   @Put(':id')
   @UsePipes(new ValidationPipe())
   updateMenuItem(@Param('id') id: string, @Body() data: Partial<MenuItemDTO>) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemService.update(id, data);
   }
 
   @Delete(':id')
   deleteMenuItem(@Param('id') id: string) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemService.delete(id);
   }
 }
